+++ ChangeLog.md
@@ -7,3 +7,12 @@
 ## Version 0.2.1 @2023/01/31
 
 * Optimize README and manual, add Change Log
+
+## Version 0.3.0 @2023/02/04
+
+Support flat format config
+
+## Version 0.3.2 @2023/02/06
+
+Global function can be a function ,in previous version only JSON is supported
+Add context to slot function